# backend/public/openapi.yaml
openapi: 3.0.3
info:
  title: SOC Platform API
  description: API documentation for the SOC Security Platform
  version: 1.0.0
servers:
  - url: https://sparky.monster
paths:
  /api/health:
    get:
      summary: Health check
      responses:
        '200':
          description: OK
  /api/alerts:
    get:
      summary: Get all security alerts
      responses:
        '200':
          description: OK
  /api/simulate-attack:
    post:
      summary: Simulate a new attack
      responses:
        '201':
          description: Created
  /api/upload:
    post:
      summary: Upload a file
      responses:
        '200':
          description: Upload successful
  /api/v3/login:
    post:
      summary: Login to get a JWT
      responses:
        '200':
          description: Returns a JWT token
  /api/v3/user/secret:
    get:
      summary: Get secure user secret
      security:
        - bearerAuth: []
      responses:
        '200':
          description: OK
  /api/v3/pet:
    post:
      summary: Add a new pet
      security:
        - bearerAuth: []
      responses:
        '201':
          description: Created
  /api/v3/pet/findByStatus:
    get:
      summary: Find pets by status
      responses:
        '200':
          description: OK
  /api/v3/pet/{id}:
    get:
      summary: Find pet by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: OK
    delete:
      summary: Delete pet by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: OK
  /api/v3/vulnerable/object/{id}:
    get:
      summary: BOLA vulnerable endpoint (Sequential ID)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: OK
  /api/v3/vulnerable/object/uuid/{uuid}:
    get:
      summary: BOLA vulnerable endpoint (UUID)
      security:
        - bearerAuth: []
      parameters:
        - name: uuid
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
  /api/v3/secure/object/{id}:
    get:
      summary: Secure object retrieval (Sequential ID)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: OK
  /api/v3/secure/object/uuid/{uuid}:
    get:
      summary: Secure object retrieval (UUID)
      security:
        - bearerAuth: []
      parameters:
        - name: uuid
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
  /api/v3/vulnerable/fake-200-auth:
    get:
      summary: Test endpoint for fake 200 OK anti-pattern
      responses:
        '200':
          description: Always returns 200 OK, but body indicates error if unauthorized
  /api/v3/graphql:
    post:
      summary: Unauthenticated GraphQL endpoint for Introspection and Depth Abuse testing
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                query:
                  type: string
                  description: The GraphQL query string
      responses:
        '200':
          description: GraphQL query result
  /api/v3/test/sensitive-error:
    get:
      summary: Demo endpoint for API security scanner validation (GET).
      tags:
        - Testing
      description: Returns simulated sensitive information in error responses for validating API security scanners.
      responses:
        '500':
          description: Internal Server Error (Simulated Leak)
          content:
            application/json:
              example:
                success: false
                message: "Unhandled exception occurred."
                stackTrace: "TypeError: Cannot read properties of undefined at UserService.js:42"
                sqlError: "SELECT * FROM users WHERE id='1' OR '1'='1'"
                internalIp: "10.0.0.15"
                awsMetadata: "169.254.169.254"
                secret: "AWS_SECRET_ACCESS_KEY=TEST_SECRET_KEY"
    post:
      summary: Demo endpoint for API security scanner validation.
      tags:
        - Testing
      description: |
        This endpoint intentionally returns simulated sensitive information in error responses for validating API security scanners. It is intended for testing environments only and must never be enabled in production.
        
        **Warning:** This endpoint is intentionally vulnerable and exists solely for QA/security scanner validation. It must be disabled or removed in production deployments.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                simulate:
                  type: boolean
                  example: true
      responses:
        '500':
          description: Internal Server Error (Simulated Leak)
          content:
            application/json:
              example:
                success: false
                message: "Unhandled exception occurred."
                stackTrace: "TypeError: Cannot read properties of undefined at UserService.js:42"
                sqlError: "SELECT * FROM users WHERE id='1' OR '1'='1'"
                database: "PostgreSQL 16"
                internalIp: "10.0.0.15"
                serverHostname: "prod-app-01"
                awsMetadata: "169.254.169.254"
                secret: "AWS_SECRET_ACCESS_KEY=TEST_SECRET_KEY"
                jwtSecret: "JWT_SECRET=test-secret-key"
                requestId: "123456789"
  /api/v3/test/mass-assignment:
    post:
      summary: Mass Assignment Vulnerability Demo
      tags:
        - Testing
      description: |
        Demo endpoint for validating Mass Assignment detection by API security scanners.
        This endpoint intentionally supports both secure and vulnerable modes.
        It exists only for QA/testing and must never be enabled in production.
      requestBody:
        required: true
        content:
          application/json:
            example:
              name: "John Doe"
              email: "john@example.com"
              isAdmin: true
              role: "admin"
              discount: 100
              tenantId: "tenant-002"
      responses:
        '201':
          description: Created object.
  /api/v3/test/mass-assignment/{id}:
    get:
      summary: Retrieve Mass Assignment Object
      tags:
        - Testing
      description: Returns the persisted object so security scanners can verify whether unauthorized properties were stored.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Retrieved object.
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT